pyproject.toml
[project]
name = "dat490"
version = "0.1.0"
description = "DAT490 Capstone: Housing-Commute Trade-Off Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "polars>=0.19.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "plotly>=5.18.0",
    "dash>=2.14.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "osmnx==2.0.6",
    "requests>=2.31.0",
    "ydata-profiling>=4.0.0",
    "prefect>=2.14.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
